// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/UserOrganisationRelationshipDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-organisation-users
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: @project.version@
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserOrganisationRelationshipDTO
 */
export interface UserOrganisationRelationshipDTO {
    /**
     * 
     * @type {string}
     * @memberof UserOrganisationRelationshipDTO
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserOrganisationRelationshipDTO
     */
    organisationName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserOrganisationRelationshipDTO
     */
    currencyCode?: string;
    /**
     * 
     * @type {string}
     * @memberof UserOrganisationRelationshipDTO
     */
    countryCode?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UserOrganisationRelationshipDTO
     */
    activated?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof UserOrganisationRelationshipDTO
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the UserOrganisationRelationshipDTO interface.
 */
export function instanceOfUserOrganisationRelationshipDTO(value: object): value is UserOrganisationRelationshipDTO {
    return true;
}

export function UserOrganisationRelationshipDTOFromJSON(json: any): UserOrganisationRelationshipDTO {
    return UserOrganisationRelationshipDTOFromJSONTyped(json, false);
}

export function UserOrganisationRelationshipDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserOrganisationRelationshipDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'organisationId': json['organisationId'] == null ? undefined : json['organisationId'],
        'organisationName': json['organisationName'] == null ? undefined : json['organisationName'],
        'currencyCode': json['currencyCode'] == null ? undefined : json['currencyCode'],
        'countryCode': json['countryCode'] == null ? undefined : json['countryCode'],
        'activated': json['activated'] == null ? undefined : json['activated'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function UserOrganisationRelationshipDTOToJSON(json: any): UserOrganisationRelationshipDTO {
    return UserOrganisationRelationshipDTOToJSONTyped(json, false);
}

export function UserOrganisationRelationshipDTOToJSONTyped(value?: UserOrganisationRelationshipDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'organisationId': value['organisationId'],
        'organisationName': value['organisationName'],
        'currencyCode': value['currencyCode'],
        'countryCode': value['countryCode'],
        'activated': value['activated'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
    };
}